change double to Bigdecimal for salary

+++ src/main/java/com/example/calculator/controller/CalculatorController.java
@@ -8,19 +8,20 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @RestController
 @RequestMapping(value = "/calculate")
 public class CalculatorController {
     @GetMapping(params = {"salary", "duration"} )
-    public double calculatePay(@RequestParam(value = "salary") @Min(0) double avgMountlySalary,
+    public BigDecimal calculatePay(@RequestParam(value = "salary") @Min(0) BigDecimal avgMountlySalary,
                                @RequestParam(value = "vacationDuration") @Min(1) int vacationDuration) {
         return Calculator.calculateByDuration(avgMountlySalary, vacationDuration);
     }
 
     @GetMapping(params = {"salary", "vacationStart", "vacationEnd"} )
-    public double calculatePay(@RequestParam(value = "salary") @Min(0) double avgMountlySalary,
+    public BigDecimal calculatePay(@RequestParam(value = "salary") @Min(0) BigDecimal avgMountlySalary,
                                @RequestParam(value = "vacationStart")  @Valid LocalDate vacationStartDate,
                                @RequestParam(value = "vacationEnd")  @Valid LocalDate vacationEndDate) {
